% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_ellipse.R
\name{geom_mvnorm_ellipse}
\alias{geom_mvnorm_ellipse}
\title{Custom Gaussian Ellipse}
\usage{
geom_mvnorm_ellipse(
  mapping = NULL,
  data = NULL,
  linewidth = 1/3,
  color = NA,
  segments = 1000,
  alpha = 1/5,
  ...,
  show.legend = NA
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes(). If specified
and inherit.aes = TRUE (the default), it is combined with the default
mapping at the top level of the plot. You must supply mapping if there
is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{linewidth}{linewidth}

\item{color}{color}

\item{segments}{I don't remember what this does}

\item{alpha}{alpha}

\item{...}{Other arguments passed on to layer()}

\item{show.legend}{self-explanatory}
}
\description{
Custom Gaussian Ellipse
}
