% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mvnorm.R
\name{geom_norm_ellipse}
\alias{geom_norm_ellipse}
\title{Title}
\usage{
geom_norm_ellipse(
  data = NULL,
  fill = "#333333",
  color = NA,
  size = 1/3,
  segments = 1000
)
}
\arguments{
\item{data}{a data frame created from a multivariate normal distribution}

\item{fill}{fill}

\item{color}{color}

\item{size}{size}

\item{segments}{I don't remember what this does}
}
\description{
Title
}
