% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{geom_rect_bkg}
\alias{geom_rect_bkg}
\title{Custom Panel Backgrounds}
\usage{
geom_rect_bkg(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes(). If specified
and inherit.aes = TRUE (the default), it is combined with the default
mapping at the top level of the plot. You must supply mapping if there
is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{...}{Other arguments passed on to layer()}
}
\description{
Custom Panel Backgrounds
}
