% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomACA-package.R
\docType{package}
\name{geomACA-package}
\alias{geomACA}
\alias{geomACA-package}
\title{geomACA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrés Castro Araújo \email{andres.castro.araujo@duke.edu}

}
\keyword{internal}
